storage_api: Avoid OverQuotaError from get_service_account_name
Commit 10e7febf00 assigned app_identity.get_service_account_name to a local
variable, outside of the if statement. This means it gets called every time
this function is called, which makes an RPC on App Engine. During periods
of high activity, this can exceed the per-minute API quota and raise
OverQuotaError.

To fix this, only call this if local_run() is True, which should only
happen in development and never in App Engine production.