Fix OauthRawGcsService::copyObject
The documentation of [x-goog-copy-source](https://cloud.google.com/storage/docs/xml-api/reference-headers#xgoogcopy)
says that the supplied path must be a legal URI and should generally be percent encoded.

x-goog-copy-source
A request header that specifies the source bucket and object for a copy operation.

Valid Values
A /<bucket>/<object> path

Implementation Details
This request header specifies the path to the source object. A valid path must include the bucket and object,
separated by a slash (/). Note that the source object path must be a legal URL, so should generally be percent encoded.